+++ tsup.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   entry: {
     index: 'src/index.ts',
     cli: 'src/cli.ts',
+    'plugins/ledger/index': 'src/plugins/ledger/index.ts',
   },
   format: ['cjs', 'esm'],
   dts: true,